requirements.txt
fastapi
uvicorn
numpy
requests
opencv-python-headless
insightface
onnxruntime-gpu
runpod
paddleocr
paddlepaddle-gpu